Updates sample exam questions
Moves the sample exam questions from a separate variable into the readings array. This change keeps the sample questions with the rest of the reading materials.

Improves the answer and explanation for sample questions 15 and 17
to provide more complete and accurate information, as well as adds a
more detailed and clear explanation for sample question 18.